refactor link retrieval logic into CRUD record handlers
